requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
idna==3.2
Markdown==3.3.4
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.26.0
short-url==1.2.2
sqlparse==0.4.1
urllib3==1.26.6
